no skipper

+++ scripts/install_environment.sh
@@ -1,6 +1,6 @@
 set -euo pipefail
 
-source utils.sh
+source scripts/utils.sh
 
 function install_libvirt() {
   if ! [ -x "$(command -v virsh)" ]; then
@@ -30,18 +30,26 @@ function install_runtime_container() {
 }
 
 function install_terraform() {
-  wget `wget https://www.terraform.io/downloads.html -q -O - | grep -oP "(https://releases.hashicorp.com/terraform/.*linux_amd64\.zip)(?=\")" | head -n 1` && unzip terraform*.zip -d /usr/bin/ && rm -rf terraform*.zip
-  wget https://dl.google.com/go/go1.13.4.linux-amd64.tar.gz && tar -C /usr/local -xf go1.13.4.linux-amd64.tar.gz && rm -f go1.13.4.linux-amd64.tar.gz
+  if ! [ -x "$(command -v terraform)" ]; then
+    wget `wget https://www.terraform.io/downloads.html -q -O - | grep -oP "(https://releases.hashicorp.com/terraform/.*linux_amd64\.zip)(?=\")" | head -n 1` && unzip terraform*.zip -d /usr/bin/ && rm -rf terraform*.zip
+  fi
+  if ! [ -x "$(command -v go)" ]; then
+    wget https://dl.google.com/go/go1.13.4.linux-amd64.tar.gz && tar -C /usr/local -xf go1.13.4.linux-amd64.tar.gz && rm -f go1.13.4.linux-amd64.tar.gz
+    echo "export PATH=/usr/local/go/bin:$PATH" >> ~/.bashrc
+    export PATH=/usr/local/go/bin:$PATH
+  fi
   mkdir -p ~/.terraform.d/plugins
   go get -v -u github.com/dmacvicar/terraform-provider-libvirt && cd ~/.terraform.d/plugins && go build -a -v github.com/dmacvicar/terraform-provider-libvirt && rm -rf ~/go
 }
 
 function install_aws() {
-    curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip" && unzip awscliv2.zip && ./aws/install && rm -rf awscliv2.zip
+  if ! [ -x "$(command -v /usr/local/bin/aws)" ]; then
+    curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip" && unzip awscliv2.zip && ./aws/install && rm -f awscliv2.zip
+  fi
 }
 
 function install_packages(){
-  dnf install -y make python3 git jq bash-completion xinetd wget unzip
+  dnf install -y make python3 git jq bash-completion xinetd wget unzip gcc
   systemctl enable --now xinetd
   pip3 install --no-cache-dir -r requirements.txt
 }
